refactor(adapterutil,rabbitmq): lift exponential backoff+jitter to shared helper
Extract addJitter/addDownJitter/backoffDelay logic from rabbitmq/connection.go
into adapterutil.ExponentialBackoffWithJitter + adapterutil.DownJitter. Delete
the empty adapters/rabbitmq/backoff.go skeleton. Update rabbitmq same-package
tests to call adapterutil helpers directly; behavior preserved verbatim.

ref: adapters/rabbitmq/connection.go reconnect wrap
ref: adapters/rabbitmq/connection.go backoffDelay
