--- adapters/rabbitmq/backoff.go
@@ -1,1 +1,0 @@
-package rabbitmq
